refactor: ttl을 long 타입으로 변경
- EX의 max 값은 9 * 10^15 정도 된다
- Int.MAX_VALUE 보다 크므로 Long으로 변경한다

+++ harrods/src/test/kotlin/com/musinsa/harrods/query/controller/QueryControllerTest.kt
@@ -23,7 +23,7 @@ class QueryControllerTest @Autowired constructor(
 ) {
     @Test
     fun `template은 빈 문자열이 아니다`() {
-        val request = MockQueryRequest(template = "", params = mapOf(), ttl = 10, interval = "* * * * *")
+        val request = MockQueryRequest(template = "", params = mapOf(), ttl = 10L, interval = "* * * * *")
 
         mvc.perform(
             post("/queries")
@@ -38,7 +38,7 @@ class QueryControllerTest @Autowired constructor(
 
     @Test
     fun `template은 null이 아니다`() {
-        val request = MockQueryRequest(template = null, params = mapOf(), ttl = 10, interval = "* * * * *")
+        val request = MockQueryRequest(template = null, params = mapOf(), ttl = 10L, interval = "* * * * *")
 
         mvc.perform(
             post("/queries")
@@ -54,7 +54,7 @@ class QueryControllerTest @Autowired constructor(
     data class MockQueryRequest(
         val template: String? = null,
         val params: Map<String, Any>? = null,
-        val ttl: Int? = null,
+        val ttl: Long? = null,
         val interval: String? = null
     )
 }
